# docs/api/openapi/hrm-api.yaml
openapi: 3.1.0
info:
  title: XeVN HRM API — Phase 1 S3b (Metadata · Spreadsheet · Operations · Performance)
  version: 1.3.0-p1-s3b
  description: |
    Canonical REST contract for HRM khối C (Sprint 3) + U18 S3b extensions.
    Scope: internal JWT or x-internal-api-key; tenant/company via headers or body.
    List paths use resolveHrmListScope — group CEO company_id=main returns rollup partition.
    ADR: docs/decisions/ADR-HRM-RBAC-SCOPE-LADDER.md
servers:
  - url: /api/hrm
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  parameters:
    CompanyIdQuery:
      name: company_id
      in: query
      required: true
      description: Company UUID or scope slug (main, holding, member slug per ADR scope ladder)
      schema: { type: string, minLength: 1, maxLength: 64 }
...
paths:
...
  /recruitment/requisitions:
    post:
      tags: [Recruitment]
      operationId: createJobRequisition
      summary: HRM-RC-01
      security: [{ bearerAuth: [] }, { internalApiKey: [] }]
      responses:
        '200':
          description: HRM-REC-201
    get:
      tags: [Recruitment]
      operationId: listJobRequisitions
      summary: HRM-RC-02
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200':
          description: HRM-REC-200
  /recruitment/candidates:
    post:
      tags: [Recruitment]
      operationId: createCandidate
      summary: HRM-RC-03
      responses:
        '200':
          description: HRM-REC-202
    get:
      tags: [Recruitment]
      operationId: listCandidates
      summary: HRM-RC-04
      parameters:
        - $ref: '#/components/parameters/CompanyIdQuery'
      responses:
        '200':
          description: HRM-REC-200
...
  /recruitment/interviews/{interviewId}/status:
    patch:
      tags: [Recruitment]
      operationId: updateInterviewStatus
      summary: HRM-RC-06
      parameters:
        - name: interviewId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: HRM-REC-204
...